requirements.txt
cmake
opencv-python
pandas
torch
torchvision
python-dotenv
pillow

fastapi-code-generator
fastapi
python-multipart
matplotlib


fastapi
uvicorn
pymongo
motor
mongoengine
pydantic
bcrypt
passlib[bcrypt]
python-jose[cryptography]
face-recognition

